GUILD-569: Add validation for producer definition to KafkaSource module
The `kafka_producer` interface is outdated at this point and it's only there
for backwards compatibility. We're moving all the validation checks into the
`kafka_producers` and throwing a deprecation warning if the old interface is used.

Incidentally, this also takes care of the infinite loop we were seeing before when
no producer method was defined.