Remove Gemfile.lock from VCS.
Gemfile.lock should not be checked-in for gem projects. This is because applications depending on the gem will rely on the dependencies specified in the .gemspec. Fixing gem versions would therefore cause the situation where a project using this gem would use a more up-to-date version of a dependency, potentially leading to compatibility issues.