Wrap `metrics` and `tracer` default values into a `default_proc`.
This is to fix an issue with the previous implementation where the initialization of the 2 mock providers would cause unwanted logs to be sent to STDOUT.
This is because the default value was calling the `#initialize` method in both `Metrics::Mock` and `Tracing::Mock` while the Deimos class was being loaded. There was therefore no way to avoid that, including specifying a custom value for these settings.